Set up Dependabot
Dependabot is a GitHub service that checks for updates to our
dependencies (in requirements.txt) and creates pull requests to update
the pinned version whenever an update is available.

For now it'll be making pull requests against master, although we should
probably make it PR against a staging branch (to test that dependency
updates don't break things) in the future.